[#5059] Cleanup ThriftProvider-Package
