+++ bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/config/DefaultProfilerConfig.java
@@ -49,10 +49,6 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     public static final int DEFAULT_AGENT_STAT_COLLECTION_INTERVAL_MS = 5 * 1000;
     public static final int DEFAULT_NUM_AGENT_STAT_BATCH_SEND = 6;
 
-    public interface ValueResolver {
-        String resolve(String value, Properties properties);
-    }
-
     private static class BypassResolver implements ValueResolver {
         public static final ValueResolver RESOLVER = new BypassResolver();
 
@@ -62,7 +58,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
         }
     }
 
-    static class PlaceHolderResolver implements ValueResolver {
+    public static class PlaceHolderResolver implements ValueResolver {
         @Override
         public String resolve(String value, Properties properties) {
             if (value == null) {
@@ -104,8 +100,6 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     private String transportModule = DEFAULT_TRANSPORT_MODULE;
 
     private ThriftTransportConfig thriftTransportConfig;
-    private GrpcTransportConfig grpcTransportConfig;
-
 
     private boolean traceAgentActiveThread = true;
 
@@ -159,7 +153,6 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     public DefaultProfilerConfig() {
         this.properties = new Properties();
         this.thriftTransportConfig = new DefaultThriftTransportConfig();
-        this.grpcTransportConfig = new GrpcTransportConfig();
     }
 
     public DefaultProfilerConfig(Properties properties) {
@@ -188,12 +181,8 @@ public class DefaultProfilerConfig implements ProfilerConfig {
         return thriftTransportConfig;
     }
 
-    public GrpcTransportConfig getGrpcTransportConfig() {
-        return grpcTransportConfig;
-    }
-
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -202,7 +191,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -211,7 +200,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -220,7 +209,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -229,7 +218,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -238,7 +227,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -247,7 +236,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -256,7 +245,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -265,7 +254,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -274,7 +263,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -283,7 +272,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -292,7 +281,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -301,7 +290,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -310,7 +299,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -319,7 +308,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -328,7 +317,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -337,7 +326,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -346,7 +335,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -355,7 +344,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -364,7 +353,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -373,7 +362,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -382,7 +371,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -391,7 +380,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -400,7 +389,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -409,7 +398,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -418,7 +407,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -427,7 +416,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -436,7 +425,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
     }
 
     /**
-     * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
+     * @deprecated Use {@link #getThriftTransportConfig()} instead.
      */
     @Deprecated
     @Override
@@ -657,8 +646,6 @@ public class DefaultProfilerConfig implements ProfilerConfig {
 
         this.transportModule = readString("profiler.transport.module", "THRIFT");
         this.thriftTransportConfig = readThriftTransportConfig(this);
-        this.grpcTransportConfig = readGrpcTransportConfig(this);
-
 
         this.traceAgentActiveThread = readBoolean("profiler.pinpoint.activethread", true);
 
@@ -733,13 +720,6 @@ public class DefaultProfilerConfig implements ProfilerConfig {
         logger.info("configuration loaded successfully.");
     }
 
-    private GrpcTransportConfig readGrpcTransportConfig(DefaultProfilerConfig profilerConfig) {
-
-        GrpcTransportConfig grpcTransportConfig = new GrpcTransportConfig();
-        grpcTransportConfig.read(profilerConfig);
-        return grpcTransportConfig;
-    }
-
     private ThriftTransportConfig readThriftTransportConfig(DefaultProfilerConfig profilerConfig) {
         DefaultThriftTransportConfig binaryTransportConfig = new DefaultThriftTransportConfig();
         binaryTransportConfig.read(profilerConfig);
@@ -752,7 +732,7 @@ public class DefaultProfilerConfig implements ProfilerConfig {
         return readString(propertyName, defaultValue, BypassResolver.RESOLVER);
     }
 
-    String readString(String propertyName, String defaultValue, ValueResolver valueResolver) {
+    public String readString(String propertyName, String defaultValue, ValueResolver valueResolver) {
         if (valueResolver == null) {
             throw new NullPointerException("valueResolver must not be null");
         }
@@ -853,7 +833,6 @@ public class DefaultProfilerConfig implements ProfilerConfig {
         sb.append(", instrumentMatcherCacheConfig=").append(instrumentMatcherCacheConfig);
         sb.append(", interceptorRegistrySize=").append(interceptorRegistrySize);
         sb.append(", thriftTransportConfig=").append(thriftTransportConfig).append('\'');
-        sb.append(", grpcTransportConfig=").append(grpcTransportConfig).append('\'');
         sb.append(", traceAgentActiveThread=").append(traceAgentActiveThread);
         sb.append(", traceAgentDataSource=").append(traceAgentDataSource);
         sb.append(", dataSourceTraceLimitSize=").append(dataSourceTraceLimitSize);
