+++ bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/config/ProfilerConfig.java
@@ -21,6 +21,7 @@ import com.navercorp.pinpoint.common.annotations.VisibleForTesting;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 
 /**
  * @author Woonduk Kang(emeroad)
@@ -32,8 +33,6 @@ public interface ProfilerConfig {
     String getTransportModule();
 
     ThriftTransportConfig getThriftTransportConfig();
-
-    GrpcTransportConfig getGrpcTransportConfig();
 
     /**
      * @deprecated Use {@link #getThriftTransportConfig()} (int)} instead.
@@ -280,6 +279,8 @@ public interface ProfilerConfig {
 
     String readString(String propertyName, String defaultValue);
 
+    String readString(String propertyName, String defaultValue, ValueResolver valueResolver);
+
     int readInt(String propertyName, int defaultValue);
 
     DumpType readDumpType(String propertyName, DumpType defaultDump);
@@ -291,5 +292,9 @@ public interface ProfilerConfig {
     boolean readBoolean(String propertyName, boolean defaultValue);
 
     Map<String, String> readPattern(String propertyNamePatternRegex);
+
+    interface ValueResolver {
+        String resolve(String value, Properties properties);
+    }
 
 }
