+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/module/GrpcModule.java
@@ -22,7 +22,7 @@ import com.google.inject.Scopes;
 import com.google.inject.TypeLiteral;
 import com.google.inject.name.Names;
 import com.google.protobuf.GeneratedMessageV3;
-import com.navercorp.pinpoint.bootstrap.config.GrpcTransportConfig;
+import com.navercorp.pinpoint.profiler.context.grpc.GrpcTransportConfig;
 import com.navercorp.pinpoint.grpc.HeaderFactory;
 import com.navercorp.pinpoint.grpc.trace.PSpan;
 import com.navercorp.pinpoint.grpc.trace.PSpanChunk;
