+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/provider/thrift/StatDataSenderProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 NAVER Corp.
+ * Copyright 2019 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.profiler.context.provider;
+package com.navercorp.pinpoint.profiler.context.provider.thrift;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import com.navercorp.pinpoint.bootstrap.config.ThriftTransportConfig;
 import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
+import com.navercorp.pinpoint.bootstrap.config.ThriftTransportConfig;
 import com.navercorp.pinpoint.common.util.Assert;
-import com.navercorp.pinpoint.profiler.context.module.SpanConverter;
 import com.navercorp.pinpoint.profiler.context.module.SpanStatClientFactory;
+import com.navercorp.pinpoint.profiler.context.module.StatConverter;
 import com.navercorp.pinpoint.profiler.context.thrift.MessageConverter;
 import com.navercorp.pinpoint.profiler.sender.DataSender;
 import com.navercorp.pinpoint.profiler.sender.MessageSerializer;
@@ -37,11 +37,11 @@ import org.slf4j.LoggerFactory;
 /**
  * @author Taejin Koo
  */
-public class SpanDataSenderProvider  implements Provider<DataSender> {
+public class StatDataSenderProvider implements Provider<DataSender> {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private static final String UDP_EXECUTOR_NAME = "Pinpoint-UdpSpanDataExecutor";
+    private static final String UDP_EXECUTOR_NAME = "Pinpoint-UdpStatDataExecutor";
 
     private final Provider<PinpointClientFactory> clientFactoryProvider;
 
@@ -52,23 +52,24 @@ public class SpanDataSenderProvider  implements Provider<DataSender> {
     private final int sendBufferSize;
     private final String ioType;
     private final String transportType;
+
     private final MessageConverter<TBase<?, ?>> messageConverter;
 
     @Inject
-    public SpanDataSenderProvider(ProfilerConfig profilerConfig, @SpanStatClientFactory Provider<PinpointClientFactory> clientFactoryProvider,
-                                  @SpanConverter MessageConverter<TBase<?, ?>> messageConverter) {
+    public StatDataSenderProvider(ProfilerConfig profilerConfig, @SpanStatClientFactory Provider<PinpointClientFactory> clientFactoryProvider, @StatConverter MessageConverter<TBase<?, ?>> messageConverter) {
         Assert.requireNonNull(profilerConfig, "profilerConfig must not be null");
+
         this.clientFactoryProvider = Assert.requireNonNull(clientFactoryProvider, "clientFactoryProvider must not be null");
 
         ThriftTransportConfig thriftTransportConfig = profilerConfig.getThriftTransportConfig();
-        this.ip = thriftTransportConfig.getCollectorSpanServerIp();
-        this.port = thriftTransportConfig.getCollectorSpanServerPort();
-        this.writeQueueSize = thriftTransportConfig.getSpanDataSenderWriteQueueSize();
-        this.timeout = thriftTransportConfig.getSpanDataSenderSocketTimeout();
-        this.sendBufferSize = thriftTransportConfig.getSpanDataSenderSocketSendBufferSize();
-        this.ioType = thriftTransportConfig.getSpanDataSenderSocketType();
-        this.transportType = thriftTransportConfig.getSpanDataSenderTransportType();
-        this.messageConverter = Assert.requireNonNull(messageConverter, "messageConverter must not be null");
+        this.ip = thriftTransportConfig.getCollectorStatServerIp();
+        this.port = thriftTransportConfig.getCollectorStatServerPort();
+        this.writeQueueSize = thriftTransportConfig.getStatDataSenderWriteQueueSize();
+        this.timeout = thriftTransportConfig.getStatDataSenderSocketTimeout();
+        this.sendBufferSize = thriftTransportConfig.getStatDataSenderSocketSendBufferSize();
+        this.ioType = thriftTransportConfig.getStatDataSenderSocketType();
+        this.transportType = thriftTransportConfig.getStatDataSenderTransportType();
+        this.messageConverter = messageConverter;
     }
 
     @Override
@@ -80,7 +81,7 @@ public class SpanDataSenderProvider  implements Provider<DataSender> {
 
             PinpointClientFactory pinpointClientFactory = clientFactoryProvider.get();
             MessageSerializer<byte[]> messageSerializer = new ThriftMessageSerializer(messageConverter);
-            return new TcpDataSender("SpanDataSender", ip, port, pinpointClientFactory, messageSerializer);
+            return new TcpDataSender("StatDataSender", ip, port, pinpointClientFactory, messageSerializer);
         } else {
             UdpDataSenderFactory factory = new UdpDataSenderFactory(ip, port, UDP_EXECUTOR_NAME, writeQueueSize, timeout, sendBufferSize, messageConverter);
             return factory.create(ioType);
@@ -89,7 +90,7 @@ public class SpanDataSenderProvider  implements Provider<DataSender> {
 
     @Override
     public String toString() {
-        final StringBuilder sb = new StringBuilder("SpanDataSenderProvider{");
+        final StringBuilder sb = new StringBuilder("StatDataSenderProvider{");
         sb.append("ip='").append(ip).append('\'');
         sb.append(", port=").append(port);
         sb.append(", writeQueueSize=").append(writeQueueSize);
